+++ package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'jam:method',
-  version: '1.7.0',
+  version: '1.7.1',
   summary: 'An easy way to create Meteor methods',
   git: 'https://github.com/jamauro/method',
   documentation: 'README.md'
@@ -11,6 +11,7 @@ Package.onUse(function(api) {
   api.use('ecmascript');
   api.use('check');
   api.use('mongo');
+  api.use('ddp-client');
   api.use('ddp-rate-limiter', 'server');
   api.use('jam:easy-schema@1.3.1', { weak: true });
   api.use('jam:offline@0.1.0', { weak: true });
@@ -23,7 +24,6 @@ Package.onTest(function(api) {
   api.use('tinytest');
   api.use('mongo');
   api.use('jam:easy-schema@1.3.1');
-  api.use('jam:offline@0.1.0');
   api.use('jam:method');
   api.addFiles('test-methods.js');
   api.mainModule('tests.js');